Moved main_sequence.py module into top-level source directory
